+++ Scripts/Extensions.cs
@@ -292,6 +292,29 @@ namespace AeLa.Utilities
 			return nearest;
 		}
 
+		public static T Peek<T>(this IList<T> list) => list[0];
+
+		public static T Pop<T>(this IList<T> list)
+		{
+			var item = list[0];
+			list.RemoveAt(0);
+			return item;
+		}
+
+		public static void Push<T>(this IList<T> list, T item) => list.Insert(0, item);
+
+		public static T Dequeue<T>(this IList<T> list)
+		{
+			var item = list[0];
+			list.RemoveAt(0);
+			return item;
+		}
+
+		public static void Enqueue<T>(this IList<T> list, T item)
+		{
+			list.Add(item);
+		}
+
 		/// <summary>
 		/// Gets the connected anchor for this joint in world space, whether or not there is a connected body.
 		/// </summary>